feat: 사용자 레벨/닉네임 커스텀 훅으로 분리
